assign and search

+++ app/Http/Controllers/Admin/ProjectUserController.php
@@ -61,8 +61,12 @@ class ProjectUserController extends Controller
                 'projects' => $projects,
                 'message' => 'Assign successfully'
             ];
-            return response()->json($data);
+        } else {
+            $data = [
+                'message' => 'Time overlaps'
+            ];
         }
+        return response()->json($data);
     }
 
     public function show($id)
@@ -107,7 +111,7 @@ class ProjectUserController extends Controller
         $data = [];
         if (is_numeric($projectId)) {
             $project = Project::with(['users' => function ($query) {
-                $query->orderBy('name');
+                $query->orderBy('name')->orderBy('start_date', 'desc');
             }])->findOrFail($projectId);
             $data = [
                 'project' => $project,
@@ -133,7 +137,7 @@ class ProjectUserController extends Controller
     public function destroyUser($userId, $projectId, $startDate, $endDate)
     {
         $project = Project::findOrFail($projectId);
-        $project->users()->wherePivot('user_id', $userId)->wherePivot('start_date', $startDate)->wherePivot('end_date', $endDate)->detach();
+        $project->users()->wherePivot('user_id', $userId)->wherePivot('start_date', $startDate)->wherePivot('end_date', $endDate)->detach($userId);
         $data = [
             'message' => 'have user',
             'startDate' => $startDate,
@@ -155,6 +159,8 @@ class ProjectUserController extends Controller
         $data2 = collect([]);
         foreach ($data1 as $user) {
             if ($startDate > $user->start_date && $endDate < $user->end_date) {
+                $data2[] = $user;
+            } elseif ($startDate < $user->start_date && $endDate > $user->end_date) {
                 $data2[] = $user;
             }
         };
